CHARLOTTE AsI Project
This is the ChatBot and Question to awnser machine I'm making called Charlotte or Cognitive Helping Artificial Responsive Linguistic Organism for Technical Task Execution

It is right now just in English. but i will translate everything to dutch as well later on. and maybe with help from other people. other languages to like Spanish, French and/or German